pyproject.toml
[project]
name = "itsm-multiagent-analysis"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
authors = [ 
    { name = "Shankar Dokula", email = "shankar.trng@gmail.com" }
]
dependencies = [
        "python-dotenv",
        "pandas",
        "numpy==2.2.5",
        "hatchling",
        "pymongo[srv]",
        "scikit-learn",
        "feast",
        "pyarrow",
        "langgraph",
        "langchain",
        "langchain-core",
        "pydantic",
        "flask",
        "langchain-community",
        "langchain-core",
        "flask-cors",

]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/itsm_analysis"]
